refactor(scheduling): 重构排课相关数据结构和逻辑
- 修改了多个类中的字段类型，从 Short 改为 Integer
- 重命名了部分类和方法，使其命名更加清晰
-优化了部分逻辑实现，提高了代码可读性和可维护性
- 调整了数据结构，使用 CourseScheduleDTO 替代原有的 ScheduleDTO

+++ src/main/java/com/frontleaves/scheduling/models/dto/scheduling/AutomaticClassSchedulingBaseDTO.java
@@ -196,17 +196,17 @@ public class AutomaticClassSchedulingBaseDTO {
              * 星期几（1-7，1表示周一）
              * 表示时间段所在的星期几。
              */
-            private Short day;
+            private Integer day;
             /**
              * 开始节次
              * 表示时间段的开始节次。
              */
-            private Short periodStart;
+            private Integer periodStart;
             /**
              * 结束节次
              * 表示时间段的结束节次。
              */
-            private Short periodEnd;
+            private Integer periodEnd;
         }
     }
 }
